+++ mythtv/programs/mythtranscode/mpeg2fix.cpp
@@ -162,7 +162,8 @@ int64_t PTSOffsetQueue::Get(int idx, AVPacket *pkt)
     while (m_offset[idx].count() > 1 && !done)
     {
         it = ++m_offset[idx].begin();
-        if ((((*it).type == 0) && (pkt->pts >= (*it).pos_pts) /* PTS type */) ||
+        if (((static_cast<int>((*it).type) == 0) &&
+             (pkt->pts >= (*it).pos_pts) /* PTS type */) ||
             (((*it).type) /* Pos type */ &&
              ((pkt->pos >= (*it).pos_pts) || (pkt->duration > (*it).framenum))))
         {
@@ -676,8 +677,8 @@ int MPEG2fixup::AddFrame(MPEG2frame *f)
         rb = &m_rx.m_vrBuf;
         rbi = &m_rx.m_indexVrbuf;
         iu.frame = GetFrameTypeN(f);
-        iu.seq_header = f->m_isSequence;
-        iu.gop = f->m_isGop;
+        iu.seq_header = static_cast<uint8_t>(f->m_isSequence);
+        iu.gop = static_cast<uint8_t>(f->m_isGop);
 
         iu.gop_off = f->m_gopPos - f->m_pkt.data;
         iu.frame_off = f->m_framePos - f->m_pkt.data;
@@ -1352,7 +1353,7 @@ int MPEG2fixup::GetFrame(AVPacket *pkt)
             m_vFrame.append(m_unreadFrames.dequeue());
             if (m_realFileEnd && !m_unreadFrames.count())
                 m_fileEnd = true;
-            return m_fileEnd;
+            return static_cast<int>(m_fileEnd);
         }
 
         while (!done)
@@ -1768,7 +1769,7 @@ int MPEG2fixup::ConvertToI(FrameList *orderedFrames, int headPos)
     foreach (auto & of, *orderedFrames)
     {
         int i = GetFrameNum(of);
-        if ((spare = DecodeToFrame(i, headPos == 0)) == nullptr)
+        if ((spare = DecodeToFrame(i, static_cast<int>(headPos == 0))) == nullptr)
         {
             LOG(VB_GENERAL, LOG_WARNING,
                 QString("ConvertToI skipping undecoded frame #%1").arg(i));
@@ -2084,7 +2085,7 @@ int MPEG2fixup::Start()
         origaPTS[it.key()] = af->first()->m_pkt.pts * 300;
         //expectedPTS[it.key()] = udiff2x33(af->first()->m_pkt.pts, initPTS);
         af_dlta_cnt[it.key()] = 0;
-        cutState[it.key()] = m_discard;
+        cutState[it.key()] = static_cast<int>(m_discard);
     }
 
     ShowRangeMap(&m_delMap, "Cutlist:");
